requirements.txt
pip3
numpy
PyQt5
paho-mqtt
requests
json